requirements.txt
# Core dependencies
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
numpy==1.23.5
scipy==1.10.1
opencv-python==4.9.0.80

# Data processing
scikit-learn==1.3.0
pandas==1.5.3
matplotlib==3.7.1
tqdm==4.65.0
Pillow==9.4.0

# Model Training
tensorboard==2.13.0

# Miscellaneous
argparse==1.4.0
